Manually exclude conflicting mixin dependencies
forgegradle uses separate dependency configurations for Minecraft
dependencies, forgeGradleMcDeps and a few others. Because configurations
are added separately to the classpath, forgegradle bypasses Gradle's
logic to evict older versions of dependencies. Gradle does eviction when
resolving a configuration, not when computing the final classpath.

Until this is fixed in forgegradle, we must manually exclude these libs.

Also exclude mixin from ClientAPI, since we depend on it here anyway.